requirements.txt
fpdf2
streamlit
watchdog
odfpy
ttkbootstrap
